fix: resolve sonar issues from protobuf migration (10 issues)
- G115: Change int32 to int64 in proto definitions and converters
- G705: Suppress XSS false positive on protojson write
- S3776: Extract helpers in generate.go and auth.go

+++ internal/orchestrator/server/convert/agent.go
@@ -37,7 +37,7 @@ func AgentDetailToProto(d *orchestrator.AgentDetails) *mobilev1.AgentDetailRespo
 		Description: d.Description,
 		AvatarUrl:   d.AvatarURL,
 		Status:      string(d.Status),
-		SortOrder:   int32(d.SortOrder),
+		SortOrder:   int64(d.SortOrder),
 		Skills:      []string{},
 		Stats: &mobilev1.AgentStatsResponse{
 			TotalMessages:         int64(d.Stats.TotalMessages),
